Reflect all iface changes and new assumptions
The new command will no longer return the same module
unless the module is calling itself (which is impossible),
so we just fall back to the dirname of the document
if nothing is found.

Also the new command no longer provides indication about
whether it's still loading anything, which is because there
is many more things that can be loaded and reloaded
and so there is practically no "finite state" worth waiting for.

Given that this "only" provides improved UI with picker dialog
and additional status bar item it seems that running it early
is an acceptable risk resulting in slightly less helpful UX.